+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.6.0
+
+* Update claim creation updates UI
+* Bug fixes and performance improvements
+* Updates inapp module dependency to 0.6.0
+
 ## 0.5.0
 
 * Add devtools ordering preference for params displayed in UI